Split `AnalogIn` into it's own module so that documentation makes a little more sese
